refactor : moved the admin job column to the table folder and the job page functions moved to helper folder and the table page heading made to reusable component
