+++ src/components/table/data-table.tsx
@@ -158,7 +158,7 @@ export function DataTable<TData, TValue>({
             ) : (
               <TableRow>
                 <TableCell colSpan={columns.length} className="h-24 text-center">
-                  No results.
+                  No Data Found In Database.
                 </TableCell>
               </TableRow>
             )}
