+++ src/router/appRouter.tsx
@@ -11,13 +11,13 @@ const Register = lazy(() => import("@/pages/auth/Register"));
 const UserJobs = lazy(() => import("@/pages/user/UserJobs"));
 const ChatPage = lazy(() => import("@/pages/common/ChatPage"));
 const Error404 = lazy(() => import("@/pages/common/Error404"));
-const AdminJobs = lazy(() => import("@/pages/admin/AdminJobs"));
 const AdminLogin = lazy(() => import("@/pages/auth/AdminLogin"));
 const AdminUsers = lazy(() => import("@/pages/admin/AdminUsers"));
 const ContactPage = lazy(() => import("@/pages/user/ContactPage"));
 const UserProfile = lazy(() => import("@/pages/user/UserProfile"));
 const AdminReports = lazy(() => import("@/pages/admin/AdminReports"));
 const AdminOverview = lazy(() => import("@/pages/admin/AdminOverview"));
+const AdminJobsPage = lazy(() => import("@/pages/admin/AdminJobsPage"));
 const AdminPackages = lazy(() => import("@/pages/admin/AdminPackages"));
 const AdminPayments = lazy(() => import("@/pages/admin/AdminPayments"));
 const AdminSettings = lazy(() => import("@/pages/admin/AdminSettings"));
@@ -127,7 +127,7 @@ const appRouter = createBrowserRouter([
                 path: "jobs",
                 element: (
                     <ProtectedRoute requiredRole={["admin", "superAdmin"]}>
-                        <AdminJobs showButton />
+                        <AdminJobsPage />
                     </ProtectedRoute>
                 ),
             },
